add testing

+++ src/test/java/ua/training/ServicesTest.java
@@ -1,9 +1,13 @@
 package ua.training;
 
 import org.junit.*;
+import ua.training.controller.command.LogoutCommand;
 import ua.training.model.entity.User;
 import ua.training.model.service.MailSender;
+import ua.training.model.service.TestDaoService;
 import ua.training.model.service.UserDaoService;
+
+import java.util.Optional;
 
 public class ServicesTest {
 
@@ -12,14 +16,23 @@ public class ServicesTest {
         UserDaoService userService = new UserDaoService();
         String login = "ivanov@test.ua";
         String password = "23456";
-        User user = userService.login(login, password);
-        Assert.assertNotNull(user.getId());
-        userService.delete(user.getId());
+        Optional<User> userOptional = userService.login(login, password);
+        if(userOptional.isPresent()){
+            User user = userOptional.get();
+            Assert.assertNotNull(user.getId());
+            userService.delete(user.getId());
+        }
     }
 
     @Test
     public void testMailSender(){
         MailSender.send("Test TESTS", "Hello! I`m testing", "test@gmail.com");
+    }
+
+    @Test
+    public void testDeteleWithAutocommit(){
+        TestDaoService testDao = new TestDaoService();
+        testDao.delete(1);
     }
 
 }
